Fixed sorting fail and crash for 1.16.1
Changed how the 'to' iterator jumps around to be a bit more logical.

Now doesnt jump back needlessly, and only moves back when about to move an item, (checking if the previous ItemStack can stack with the held one, and isnt full), in which case it will fill there instead, before moving on.